+++ DependencyInjection/SpBowerExtension.php
@@ -67,8 +67,16 @@ class SpBowerExtension extends Extension
             $execListener->addTag('kernel.event_subscriber');
         }
 
-        $container->setParameter('sp_bower.bower.offline', $config['offline']);
-        $container->setParameter('sp_bower.bower.allow_root', $config['allow_root']);
+        $options = $config['options'];
+        if (!isset($options['offline']) && $config['offline']) {
+            $options['offline'] = $config['offline'];
+        }
+
+        if (!isset($options['allow-root']) && $config['allow_root']) {
+            $options['allow-root'] = $config['allow_root'];
+        }
+
+        $container->setParameter('sp_bower.bower.options', $options);
         $container->setParameter('sp_bower.bower.bin', $config['bin']);
         $container->setParameter('sp_bower.install_on_warmup', $config['install_on_warmup']);
         $this->loadBundlesInformation($config['bundles'], $container);